Using/packaging XULRunner. Cleaned up (un)install.
Apps now run on XULRunner, which is currently packaged up in the extension.
We now write install.exe to a temporary directory so it doesn't clutter the
installation dir.  We now install to %APPDATA% instead of "%LOCALAPPDATA%\Web
Apps".  Each app now gets its own uninstaller.  We now remove the local
storage associated with an app on uninstall.