Removing XULRunner, supporting data URLs, cleanup partial installs.
Creating shortcuts instead of a native binary; this lets us remove (the
Windows version of) XULRunner from the extension. Moved "get icon" code into
a shared function.  This way Mac and Windows support for icon data URLs and
non-data URLs can be shared.  I think I broke Mac icon synthesis in the
process (sorry Dan!).  Removing files and attempting to run uninstaller on
failed install attempts.  Using a .ini file to send arguments to installer
instead of command line args.  This avoids issues with quotes in app names,
though we still need to sanitize the app name before writing it to the file
(will do this in a later commit).