Various Windows-specific extension fixes.
Bug 704616 - Using app origin instead of app name in uninstall metadata.  This avoids accidentally trashing uninstall metadata for other apps with the same name.  Also installing to %APPDATA%/Mozilla instead of %APPDATA% to avoid overwriting the directories of other apps that live in %APPDATA% (e.g. if a webapp decides to call itself "Dropbox" we no longer trash users' Dropbox installations).

Bug 704566 - Initial support for app descriptions.  We should add a separate
field in the manifest for this (see discussion in bug), but for now we use the "description" field as a proof of concept.

Bug 701609 - Cleaning up installer after installation (otherwise we clutter the user's temp directory with installers).

Bug 701610 - Adding install.exe back to the repo