bug 701609 - Fix partial install issues.
If the uninstaller is present, rely on it to remove the installation dir. Fixes issue where we inadvertently removed the installation dir out from under the uninstaller, so it failed to run properly. In the callback for icon retrieval, do not proceed if the installation has already aborted, and attemp to remove the installation;r=anant